Implement new error handling for jst generators
